change_props_data
